Rename imitation.scripts.common to imitation.scripts.ingredients
