+++ src/imitation/scripts/config/train_preference_comparisons.py
@@ -3,7 +3,7 @@
 import sacred
 
 from imitation.algorithms import preference_comparisons
-from imitation.scripts.common import common, reward, rl, train
+from imitation.scripts.ingredients import common, reward, rl, train
 
 train_preference_comparisons_ex = sacred.Experiment(
     "train_preference_comparisons",
