+++ src/imitation/scripts/ingredients/rl.py
@@ -17,8 +17,8 @@ from stable_baselines3.common import (
 from imitation.policies import serialize
 from imitation.policies.replay_buffer_wrapper import ReplayBufferRewardWrapper
 from imitation.rewards.reward_function import RewardFn
-from imitation.scripts.common import common
-from imitation.scripts.common.train import train_ingredient
+from imitation.scripts.ingredients import common
+from imitation.scripts.ingredients.train import train_ingredient
 
 rl_ingredient = sacred.Ingredient(
     "rl",
