+++ src/imitation/scripts/train_preference_comparisons.py
@@ -15,12 +15,12 @@ from stable_baselines3.common import type_aliases
 from imitation.algorithms import preference_comparisons
 from imitation.data import types
 from imitation.policies import serialize
-from imitation.scripts.common import common, reward
-from imitation.scripts.common import rl as rl_common
-from imitation.scripts.common import train
 from imitation.scripts.config.train_preference_comparisons import (
     train_preference_comparisons_ex,
 )
+from imitation.scripts.ingredients import common, reward
+from imitation.scripts.ingredients import rl as rl_common
+from imitation.scripts.ingredients import train
 
 
 def save_model(
